refactor: cmd/serverの処理をhandlerに切り離した
