+++ backend/cmd/server/main.go
@@ -3,24 +3,14 @@ package main
 import (
 	"database/sql"
 	"fmt"
-	"net/http"
 	"os"
-	"strconv"
-	"time"
 
 	"github.com/go-sql-driver/mysql"
-	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
+	"github.com/saitamau-maximum/maxitter/backend/handler"
 	"github.com/uptrace/bun"
 	"github.com/uptrace/bun/dialect/mysqldialect"
-
-	"github.com/saitamau-maximum/maxitter/backend/model"
 )
-
-type Handler struct {
-	DB     *bun.DB
-	Logger echo.Logger
-}
 
 var (
 	SQL_PATH   = "./sql"
@@ -74,7 +64,7 @@ func main() {
 	bunDB := bun.NewDB(db, mysqldialect.New())
 	defer bunDB.Close()
 
-	h := &Handler{DB: bunDB, Logger: e.Logger}
+	h := &handler.Handler{DB: bunDB, Logger: e.Logger}
 	api := e.Group("/api")
 	api.GET("/posts", h.GetPosts)
 	api.POST("/posts", h.CreatePost)
@@ -85,103 +75,4 @@ func main() {
 	api.GET("/users", h.GetUsers)
 	api.POST("/users/new", h.CreateUser)
 	e.Logger.Fatal(e.Start(":8000"))
-}
-
-func (h *Handler) GetPosts(c echo.Context) error {
-	pageParam := c.QueryParam("page")
-	if pageParam == "" {
-		pageParam = "0"
-	}
-	page, err := strconv.ParseUint(pageParam, 10, 0)
-	if err != nil {
-		return c.JSON(400, err)
-	}
-
-	index := page * 20
-
-	ctx := c.Request().Context()
-
-	var modelPosts []model.Post
-	err = h.DB.NewSelect().Model(&modelPosts).Order("created_at DESC").Limit(20).Offset(int(index)).Scan(ctx)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	return c.JSON(http.StatusOK, modelPosts)
-}
-
-func (h *Handler) CreatePost(c echo.Context) error {
-	id, err := uuid.NewRandom()
-	if err != nil {
-		h.Logger.Error(err)
-		return c.JSON(500, err)
-	}
-	post := new(model.Post)
-	if err := c.Bind(post); err != nil {
-		h.Logger.Error(err)
-		return c.JSON(500, err)
-	}
-	post.ID = id.String()
-	post.CreatedAt = time.Now().Round(time.Millisecond)
-
-	modelPost := &model.Post{
-		ID:        post.ID,
-		Body:      post.Body,
-		CreatedAt: post.CreatedAt,
-	}
-
-	ctx := c.Request().Context()
-
-	_, err = h.DB.NewInsert().Model(modelPost).Exec(ctx)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-	return c.JSON(http.StatusOK, modelPost)
-}
-
-func (h *Handler) GetUsers(c echo.Context) error {
-	ctx := c.Request().Context()
-
-	var modelUsers []model.User
-	err := h.DB.NewSelect().Model(&modelUsers).Scan(ctx)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	return c.JSON(http.StatusOK, modelUsers)
-}
-
-func (h *Handler) CreateUser(c echo.Context) error {
-	id, err := uuid.NewRandom()
-	if err != nil {
-		h.Logger.Error(err)
-		return c.JSON(500, err)
-	}
-	user := new(model.User)
-	if err := c.Bind(user); err != nil {
-		h.Logger.Error(err)
-		return c.JSON(500, err)
-	}
-	user.ID = id.String()
-
-	user.CreatedAt = time.Now().Round(time.Millisecond)
-	user.UpdatedAt = time.Now().Round(time.Millisecond)
-
-	modelUser := &model.User{
-		ID:           user.ID,
-		Name:         user.Name,
-		ProfileImage: user.ProfileImage,
-		Bio:          user.Bio,
-		CreatedAt:    user.CreatedAt,
-		UpdatedAt:    user.UpdatedAt,
-	}
-
-	ctx := c.Request().Context()
-
-	_, err = h.DB.NewInsert().Model(modelUser).Exec(ctx)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, err.Error())
-	}
-
-	return c.JSON(http.StatusOK, modelUser)
 }
